Refactor crank build commands to build static packages
Refactors crank build <package-type> commands to build static packages
from a scratch image using go-containerregistry. These packages are
single layer with the layer being a tarball containing only a
package.yaml, which has the contents of the package in a YAML stream.

Note that the package images are saved with the .xpkg extension, which
identifies them as Crossplane packages. Crossplane packages can be
loaded into a container runtime, such as Docker, just as any OCI image
tarball can. Importantly, Crossplane package images are not tagged on
build, but are instead tagged on push before being written to a
registry.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>