Use our own Reference types
Previously we used *xpv1.Reference (from crossplane-runtime) in a few
places in this codebase. This type was recently updated to add a
'policy' field that isn't applicable to core Crossplane's uses.

Reference structs are simple enough that it seemed better to define our
own than to depend on (and have to name!) a new kind of reference in
crossplane-runtime.

Signed-off-by: Nic Cope <nicc@rk0n.org>