render: Try harder to clean up when render times out
Previously, the cleanup step in render (which removes function containers
started by the render process) was using the context passed to the main render
operation. This meant that if the render timed out the cleanup would be skipped,
since the context was already canceled.

Use a separate context for cleanup with a relatively short timeout (5s) so that
cleanup can proceed in timeout conditions.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>